+++ pyquil/api/_config.py
@@ -158,9 +158,6 @@ class PyquilConfig(object):
         # The engagement callback can be added by config consumers after construction
         self.get_engagement: Optional[Callable[..., Optional[Engagement]]] = lambda: None
 
-        # Whether engagement has been requested in order to provide any config values
-        self._engagement_requested = False
-
         self.config_parsers: Dict[str, ConfigParser] = {}
         for env_name, default_path in config_paths.items():
             default_path = expanduser(default_path)
@@ -218,7 +215,6 @@ class PyquilConfig(object):
         assert self.get_engagement is not None
         try:
             if engagement_key is not None and self.get_engagement() is not None:
-                self._engagement_requested = True
                 return cast(str, getattr(self.get_engagement(), engagement_key))
         except AttributeError:
             pass
@@ -241,17 +237,6 @@ class PyquilConfig(object):
     @property
     def engage_cmd(self) -> Optional[str]:
         return self._env_or_config_or_default(**self.ENGAGE_CMD)
-
-    @property
-    def engagement(self) -> Optional["Engagement"]:
-        """
-        An Engagement should only be made available to consumers if it was used to retrieve a
-        configuration value.
-        """
-        if not self._engagement_requested:
-            return None
-        assert self.get_engagement is not None
-        return self.get_engagement()
 
     @property
     def forest_url(self) -> Optional[str]: